Add files via upload
8. Process Odd Positions
You are given an array of numbers. Write a JS function that returns the elements at odd positions from the array, doubled and in reverse order.
The input comes as an array of number elements.
The output is the return on the console on a single line, separated by space.