Add files via upload
11. Equal Neighbors
Write a function that finds the number of equal neighbor pairs inside a matrix of variable size and type (numbers or strings).
The input comes as an array of arrays, containing string elements (2D matrix of strings).
The output is the return value of your function. Save the number of equal pairs you find and return it.